Making the input and the system state exclusive for each test case eliminates the dependency among the test cases.
